modified the way of detecting local IP address.
added getLocalIP() method in the CMCommManager class.
modified init() method in the CMConfigurator class.
In order to set my address in the CMConfigurator, we call
CMCommManager.getLocalIP() method.
To detect local IP, we detect inet addresses bound to the local host,
and find the local IP instead of simply calling
InetAddress.getLocalHost().getHostAddress() method, which has a problem
of incorrect detection in some network environments.