created CMMqttEventSUBSCRIBE class.
added member variables and methods.
found a bug while testing the CMMqttEventSUBSCRIBE in the CMMqttEventTest class.
The problem is about the conflict of string marshalling/unmarshalling in the CM event header and CM mqtt event body.
We need to modify marshalling/unmarshalling of the existing CM events such that the string length preceding string bytes should be changed to 2 bytes (MQTT style) instead of 4 bytes (existing CM style).

수정함:        CM/src/CMMqttEventTest.java
새 파일:       CM/src/kr/ac/konkuk/ccslab/cm/event/mqttevent/CMMqttEventSUBSCRIBE.java